+++ src/types/comfyManagerTypes.ts
